refactor(cli): split cli_show_services.go by service family (#4655)
pkg/cli/cli_show_services.go had grown to 868 LOC and mixed the
presenters for unrelated service families — class-of-service, dynamic
DNS, DHCP (client/relay/server), SNMP, LLDP, and port mirroring — in one
file. Different owners editing different services kept touching the same
bucket, and the file was a merge-conflict magnet.

Split the presenters into six sibling files grouped by natural service
family, keeping cli_show_services.go as the `show services` dispatch
shell plus the presenters that have no dedicated family file (RPM,
ip-monitoring, application-identification, schedulers):

  show_services_cos.go    handleShowClassOfService, parseCoSClassifierArgs,
                          showClassOfServiceInterface, showInterfacesQueue
  show_services_ddns.go   showServicesDynamicDNS, showDHCPDynamicDNS
  show_services_dhcp.go   showDHCPLeases, showDHCPClientIdentifier,
                          showDHCPRelay, showDHCPServer
  show_services_snmp.go   showSNMP, showSNMPv3
  show_services_lldp.go   showLLDP, showLLDPNeighbors
  show_services_mirror.go showPortMirroring

This is a pure code-motion refactor with no behavior change. All 20
top-level declarations were moved verbatim; the files stay in package
`cli`, so the unexported helpers (showConfigRedacted,
userspaceDataplaneStatus) and the injected `*Fn` status hooks remain
reachable. Per-file import lists were narrowed to what each file uses.

Validation: a go/parser decl-diff confirms every top-level declaration
is byte-for-byte identical between origin/master's single file and the
union of the new files (identical=20, 0 bodydiff / 0 missing / 0 extra;
import declarations legitimately differ per file and are excluded).
`go build ./...` clean, `go test ./pkg/cli/...` green, gofmt clean. The
one `go vet` note (cli.go:503 unreachable code) pre-dates this change.
Docs updated: pkg/cli/README.md gains a "show services presenters"
layout section, and the docs/system-login.md pointer to `showSNMP` now
names show_services_snmp.go.

Closes #4655
