+++ pkg/cli/cli_show_services.go
@@ -3,74 +3,13 @@ package cli
 import (
 	"fmt"
 	"os"
-	"sort"
 	"strings"
-	"time"
 
 	"github.com/psaab/xpf/pkg/appid"
 	"github.com/psaab/xpf/pkg/cmdtree"
-	"github.com/psaab/xpf/pkg/config"
-	dpuserspace "github.com/psaab/xpf/pkg/dataplane/userspace"
-	dpformat "github.com/psaab/xpf/pkg/dataplane/userspace/format"
-	"github.com/psaab/xpf/pkg/dhcp"
-	"github.com/psaab/xpf/pkg/dhcpserver"
 	"github.com/psaab/xpf/pkg/ipmon"
 	"github.com/psaab/xpf/pkg/rpm"
 )
-
-// handleShowClassOfService dispatches `show class-of-service ...`
-// to the appropriate presenter.
-func (c *CLI) handleShowClassOfService(args []string) error {
-	if len(args) == 0 {
-		cmdtree.PrintTreeHelp("show class-of-service:", operationalTree, "show", "class-of-service")
-		return nil
-	}
-	switch args[0] {
-	case "interface":
-		selector := ""
-		if len(args) > 1 {
-			selector = args[1]
-		}
-		return c.showClassOfServiceInterface(selector)
-	case "classifier":
-		nameFilter, typeFilter := parseCoSClassifierArgs(args[1:])
-		fmt.Print(dpformat.FormatCoSClassifiers(c.store.ActiveConfig(), nameFilter, typeFilter))
-		return nil
-	case "scheduler-map":
-		name := ""
-		if len(args) > 1 {
-			name = args[1]
-		}
-		fmt.Print(dpformat.FormatCoSSchedulerMaps(c.store.ActiveConfig(), name))
-		return nil
-	case "forwarding-class":
-		fmt.Print(dpformat.FormatCoSForwardingClasses(c.store.ActiveConfig()))
-		return nil
-	default:
-		cmdtree.PrintTreeHelp("show class-of-service:", operationalTree, "show", "class-of-service")
-		return nil
-	}
-}
-
-// parseCoSClassifierArgs extracts optional `name <n>` and `type <t>` filters
-// from `show class-of-service classifier` arguments.
-func parseCoSClassifierArgs(args []string) (nameFilter, typeFilter string) {
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "name":
-			if i+1 < len(args) {
-				nameFilter = args[i+1]
-				i++
-			}
-		case "type":
-			if i+1 < len(args) {
-				typeFilter = args[i+1]
-				i++
-			}
-		}
-	}
-	return nameFilter, typeFilter
-}
 
 // handleShowServices dispatches `show services ...` to the appropriate
 // presenter (RPM, application-identification, ...).
@@ -125,197 +64,6 @@ func (c *CLI) handleShowServices(args []string) error {
 	default:
 		return fmt.Errorf("unknown services target: %s", args[0])
 	}
-}
-
-// showServicesDynamicDNS renders the Surface A (router/interface-address) DDNS
-// status: the provider catalog, runtime counters, and (detail) the per-scope
-// last-published state (#2691 P2).
-func (c *CLI) showServicesDynamicDNS(detail bool) error {
-	cfg := c.store.ActiveConfig()
-	fmt.Println("Dynamic DNS (Surface A — router/interface-address publish):")
-	if cfg != nil && cfg.System.Services != nil && cfg.System.Services.DynamicDNS != nil {
-		cat := cfg.System.Services.DynamicDNS
-		if len(cat.Providers) > 0 {
-			fmt.Println("  Providers:")
-			names := make([]string, 0, len(cat.Providers))
-			for n := range cat.Providers {
-				names = append(names, n)
-			}
-			sort.Strings(names)
-			for _, n := range names {
-				p := cat.Providers[n]
-				backend := p.Backend
-				if backend == "" {
-					backend = "rfc2136"
-				}
-				fmt.Printf("    %s: backend=%s update-server=%s", n, backend, p.UpdateServer)
-				if p.TSIGKeyName != "" {
-					fmt.Printf(" tsig-key=%s (secret redacted)", p.TSIGKeyName)
-				}
-				fmt.Println()
-			}
-		}
-		if cat.ForcedRefreshSeconds > 0 {
-			fmt.Printf("  Forced-refresh: %ds\n", cat.ForcedRefreshSeconds)
-		}
-		if cat.ErrorBackoffMaxSeconds > 0 {
-			fmt.Printf("  Error-backoff-max: %ds\n", cat.ErrorBackoffMaxSeconds)
-		}
-	} else {
-		fmt.Println("  No provider catalog configured")
-	}
-
-	if c.surfaceADDNSStatsFn == nil {
-		fmt.Println("\n  Runtime counters: unavailable")
-		return nil
-	}
-	st := c.surfaceADDNSStatsFn()
-	if st == nil {
-		fmt.Println("\n  Runtime counters: unavailable (manager not running)")
-		return nil
-	}
-	if st.Degraded {
-		fmt.Printf("\n  ALARM: Surface A DDNS DEGRADED (fail-closed) — %s\n", st.DegradedReason)
-		fmt.Println("    Publishing and withdrawals are SUSPENDED until the ownership state is resolved.")
-	}
-	fmt.Println("\n  Counters:")
-	fmt.Printf("    Publishes: ok=%d fail=%d\n", st.UpsertOK, st.UpsertFail)
-	fmt.Printf("    Withdraws: ok=%d fail=%d\n", st.DeleteOK, st.DeleteFail)
-	fmt.Printf("    Skipped:   unchanged=%d backoff=%d no-backend=%d\n", st.Skipped, st.BackedOff, st.SkippedNoBackend)
-	fmt.Printf("    Published records: %d\n", st.Scopes)
-
-	if detail && c.surfaceADDNSStatusFn != nil {
-		views := c.surfaceADDNSStatusFn()
-		fmt.Println("\n  Configured scopes:")
-		if len(views) == 0 {
-			fmt.Println("    none")
-		} else {
-			fmt.Printf("    %-32s %-6s %-15s %-39s %-12s %s\n", "FQDN", "Family", "State", "Address", "Provider", "Last error")
-			for _, v := range views {
-				fam := "inet"
-				if v.Family == 6 {
-					fam = "inet6"
-				}
-				lastErr := v.LastError
-				if lastErr == "" {
-					lastErr = "-"
-				}
-				addr := v.Published
-				if addr == "" {
-					addr = "-"
-				}
-				fmt.Printf("    %-32s %-6s %-15s %-39s %-12s %s\n",
-					v.FQDN, fam, v.State, addr, v.Provider, lastErr)
-			}
-		}
-	}
-	return nil
-}
-
-func (c *CLI) showDHCPLeases() error {
-	if c.dhcp == nil {
-		fmt.Println("No DHCP clients running")
-		return nil
-	}
-
-	leases := c.dhcp.Leases()
-	if len(leases) == 0 {
-		fmt.Println("No active DHCP leases")
-		return nil
-	}
-
-	fmt.Println("DHCP leases:")
-	for _, l := range leases {
-		family := "inet"
-		if l.Family == dhcp.AFInet6 {
-			family = "inet6"
-		}
-		elapsed := time.Since(l.Obtained).Round(time.Second)
-		remaining := l.LeaseTime - elapsed
-		if remaining < 0 {
-			remaining = 0
-		}
-		fmt.Printf("  Interface: %s, Family: %s\n", l.Interface, family)
-		fmt.Printf("    Address:   %s\n", l.Address)
-		if l.Gateway.IsValid() {
-			fmt.Printf("    Gateway:   %s\n", l.Gateway)
-		}
-		if len(l.DNS) > 0 {
-			dnsStrs := make([]string, len(l.DNS))
-			for i, d := range l.DNS {
-				dnsStrs[i] = d.String()
-			}
-			fmt.Printf("    DNS:       %s\n", strings.Join(dnsStrs, ", "))
-		}
-		fmt.Printf("    Lease:     %s (remaining: %s)\n", l.LeaseTime.Round(time.Second), remaining.Round(time.Second))
-		fmt.Printf("    Obtained:  %s\n", l.Obtained.Format("2006-01-02 15:04:05"))
-		fmt.Println()
-	}
-
-	// Show delegated prefixes
-	pds := c.dhcp.DelegatedPrefixes()
-	if len(pds) > 0 {
-		fmt.Println("Delegated prefixes (DHCPv6 PD):")
-		for _, dp := range pds {
-			elapsed := time.Since(dp.Obtained).Round(time.Second)
-			remaining := dp.ValidLifetime - elapsed
-			if remaining < 0 {
-				remaining = 0
-			}
-			fmt.Printf("  Interface: %s\n", dp.Interface)
-			fmt.Printf("    Prefix:    %s\n", dp.Prefix)
-			fmt.Printf("    Preferred: %s\n", dp.PreferredLifetime.Round(time.Second))
-			fmt.Printf("    Valid:     %s (remaining: %s)\n", dp.ValidLifetime.Round(time.Second), remaining.Round(time.Second))
-			fmt.Printf("    Obtained:  %s\n", dp.Obtained.Format("2006-01-02 15:04:05"))
-			fmt.Println()
-		}
-	}
-
-	return nil
-}
-
-func (c *CLI) showDHCPClientIdentifier() error {
-	if c.dhcp == nil {
-		fmt.Println("No DHCP clients running")
-		return nil
-	}
-
-	duids := c.dhcp.DUIDs()
-	if len(duids) == 0 {
-		fmt.Println("No DHCPv6 DUIDs configured")
-		return nil
-	}
-
-	fmt.Println("DHCPv6 client identifiers:")
-	for _, d := range duids {
-		fmt.Printf("  Interface: %s\n", d.Interface)
-		fmt.Printf("    Type:    %s\n", d.Type)
-		fmt.Printf("    DUID:    %s\n", d.Display)
-		fmt.Printf("    Hex:     %s\n", d.HexBytes)
-		fmt.Println()
-	}
-	return nil
-}
-
-func (c *CLI) showClassOfServiceInterface(selector string) error {
-	cfg := c.store.ActiveConfig()
-	var status *dpuserspace.ProcessStatus
-	if userspaceStatus, err := c.userspaceDataplaneStatus(); err == nil {
-		status = &userspaceStatus
-	}
-	fmt.Print(dpformat.FormatCoSInterfaceSummary(cfg, status, selector))
-	return nil
-}
-
-// showInterfacesQueue renders `show interfaces queue [<interface>]` (#4228
-// Gap 7) from the live userspace CoS runtime snapshot.
-func (c *CLI) showInterfacesQueue(selector string) error {
-	var status *dpuserspace.ProcessStatus
-	if userspaceStatus, err := c.userspaceDataplaneStatus(); err == nil {
-		status = &userspaceStatus
-	}
-	fmt.Print(dpformat.FormatInterfacesQueue(status, selector))
-	return nil
 }
 
 func (c *CLI) showRPMProbeResults() error {
@@ -421,446 +169,6 @@ func (c *CLI) showSchedulers() error {
 		}
 		if sched.Daily {
 			fmt.Println("  Recurrence: daily")
-		}
-		fmt.Println()
-	}
-	return nil
-}
-
-func (c *CLI) showDHCPRelay() error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil || cfg.ForwardingOptions.DHCPRelay == nil {
-		fmt.Println("No DHCP relay configured")
-		return nil
-	}
-	relay := cfg.ForwardingOptions.DHCPRelay
-
-	if len(relay.ServerGroups) > 0 {
-		fmt.Println("Server groups:")
-		for name, sg := range relay.ServerGroups {
-			fmt.Printf("  %s: %s\n", name, strings.Join(sg.Servers, ", "))
-		}
-	}
-
-	if len(relay.Groups) > 0 {
-		fmt.Println("Relay groups:")
-		for name, g := range relay.Groups {
-			fmt.Printf("  %s:\n", name)
-			fmt.Printf("    Interfaces: %s\n", strings.Join(g.Interfaces, ", "))
-			fmt.Printf("    Active server group: %s\n", g.ActiveServerGroup)
-			var overrides []string
-			if g.AlwaysBroadcast {
-				overrides = append(overrides, "always-broadcast")
-			}
-			// #4309: maximum-hop-count is enforced; forward-only /
-			// relay-agent-option are accepted-only (annotated so the operator
-			// sees they match the relay's existing default behavior).
-			if g.MaximumHopCount > 0 {
-				overrides = append(overrides, fmt.Sprintf("maximum-hop-count %d", g.MaximumHopCount))
-			}
-			if g.ForwardOnly {
-				overrides = append(overrides, "forward-only (accepted-only)")
-			}
-			if g.RelayAgentOption {
-				overrides = append(overrides, "relay-agent-option (accepted-only)")
-			}
-			if len(overrides) > 0 {
-				fmt.Printf("    Overrides: %s\n", strings.Join(overrides, ", "))
-			}
-		}
-	}
-
-	// Runtime statistics
-	if c.dhcpRelay != nil {
-		stats := c.dhcpRelay.Stats()
-		if len(stats) > 0 {
-			fmt.Println("\nRelay statistics:")
-			fmt.Printf("  %-16s %-20s %-20s %s\n", "Interface", "Requests relayed", "Replies forwarded", "Dropped (max-hops)")
-			for _, s := range stats {
-				fmt.Printf("  %-16s %-20d %-20d %d\n", s.Interface, s.RequestsRelayed, s.RepliesForwarded, s.RequestsDroppedMaxHops)
-			}
-			// Reply-delivery breakdown (#2076). L2-fallback is the one to
-			// alert on: it means the raw-L2 path failed (CAP_NET_RAW,
-			// driver, or MTU) and the relay degraded to broadcast.
-			fmt.Println("\nReply delivery (#2076):")
-			fmt.Printf("  %-16s %-10s %-10s %-10s %-10s %-10s %-12s %s\n",
-				"Interface", "L2-unicast", "ciaddr", "bcast-flag", "bcast-fwd",
-				"no-target", "L2-fallback", "nak-bcast")
-			for _, s := range stats {
-				fmt.Printf("  %-16s %-10d %-10d %-10d %-10d %-10d %-12d %d\n",
-					s.Interface, s.RepliesL2Unicast, s.RepliesUnicastCiaddr,
-					s.RepliesBroadcastFlag1, s.RepliesBroadcastForced,
-					s.RepliesBroadcastNoTarget, s.RepliesBroadcastL2Fallback,
-					s.RepliesBroadcastNak)
-			}
-			// Reply source validation (#4163). A non-zero count means a reply
-			// arrived from a source IP that is NOT one of the configured DHCP
-			// servers and was dropped — a rogue-reply injection attempt, or a
-			// multi-homed server unicasting from an unlisted source IP.
-			fmt.Println("\nReply source validation (#4163):")
-			fmt.Printf("  %-16s %s\n", "Interface", "Dropped (unknown server)")
-			for _, s := range stats {
-				fmt.Printf("  %-16s %d\n", s.Interface, s.RepliesDroppedUnknownServer)
-			}
-		}
-	}
-	return nil
-}
-
-func (c *CLI) showDHCPServer(detail bool) error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil || (cfg.System.DHCPServer.DHCPLocalServer == nil && cfg.System.DHCPServer.DHCPv6LocalServer == nil) {
-		fmt.Println("No DHCP server configured")
-		return nil
-	}
-
-	// In detail mode, show pool configuration first
-	if detail {
-		if srv := cfg.System.DHCPServer.DHCPLocalServer; srv != nil && len(srv.Groups) > 0 {
-			fmt.Println("DHCPv4 Server Configuration:")
-			for name, group := range srv.Groups {
-				fmt.Printf("  Group: %s\n", name)
-				if len(group.Interfaces) > 0 {
-					fmt.Printf("    Interfaces: %s\n", strings.Join(group.Interfaces, ", "))
-				}
-				for _, pool := range group.Pools {
-					fmt.Printf("    Pool: %s\n", pool.Name)
-					if pool.Subnet != "" {
-						fmt.Printf("      Subnet: %s\n", pool.Subnet)
-					}
-					if pool.RangeLow != "" {
-						fmt.Printf("      Range: %s - %s\n", pool.RangeLow, pool.RangeHigh)
-					}
-					if pool.Router != "" {
-						fmt.Printf("      Router: %s\n", pool.Router)
-					}
-					if len(pool.DNSServers) > 0 {
-						fmt.Printf("      DNS: %s\n", strings.Join(pool.DNSServers, ", "))
-					}
-					if pool.LeaseTime > 0 {
-						fmt.Printf("      Lease time: %ds\n", pool.LeaseTime)
-					}
-				}
-			}
-			fmt.Println()
-		}
-		if srv := cfg.System.DHCPServer.DHCPv6LocalServer; srv != nil && len(srv.Groups) > 0 {
-			fmt.Println("DHCPv6 Server Configuration:")
-			for name, group := range srv.Groups {
-				fmt.Printf("  Group: %s\n", name)
-				if len(group.Interfaces) > 0 {
-					fmt.Printf("    Interfaces: %s\n", strings.Join(group.Interfaces, ", "))
-				}
-				for _, pool := range group.Pools {
-					fmt.Printf("    Pool: %s\n", pool.Name)
-					if pool.Subnet != "" {
-						fmt.Printf("      Subnet: %s\n", pool.Subnet)
-					}
-					if pool.RangeLow != "" {
-						fmt.Printf("      Range: %s - %s\n", pool.RangeLow, pool.RangeHigh)
-					}
-				}
-			}
-			fmt.Println()
-		}
-	}
-
-	// Read Kea lease files directly.
-	server := dhcpserver.New()
-	leases4, _ := server.GetLeases4()
-	leases6, _ := server.GetLeases6()
-
-	if len(leases4) == 0 && len(leases6) == 0 {
-		if !detail {
-			fmt.Println("No active leases")
-		} else {
-			fmt.Println("Active leases: none")
-		}
-		return nil
-	}
-
-	if len(leases4) > 0 {
-		fmt.Printf("DHCPv4 Leases (%d active):\n", len(leases4))
-		if detail {
-			fmt.Printf("  %-18s %-20s %-15s %-10s %-12s %s\n", "Address", "MAC", "Hostname", "Subnet", "Lifetime", "Expires")
-			for _, l := range leases4 {
-				fmt.Printf("  %-18s %-20s %-15s %-10s %-12s %s\n",
-					l.Address, l.HWAddress, l.Hostname, l.SubnetID, l.ValidLife, l.ExpireTime)
-			}
-		} else {
-			fmt.Printf("  %-18s %-20s %-15s %-12s %s\n", "Address", "MAC", "Hostname", "Lifetime", "Expires")
-			for _, l := range leases4 {
-				fmt.Printf("  %-18s %-20s %-15s %-12s %s\n",
-					l.Address, l.HWAddress, l.Hostname, l.ValidLife, l.ExpireTime)
-			}
-		}
-	}
-	if len(leases6) > 0 {
-		fmt.Printf("DHCPv6 Leases (%d active):\n", len(leases6))
-		if detail {
-			fmt.Printf("  %-40s %-20s %-15s %-10s %-12s %s\n", "Address", "DUID", "Hostname", "Subnet", "Lifetime", "Expires")
-			for _, l := range leases6 {
-				fmt.Printf("  %-40s %-20s %-15s %-10s %-12s %s\n",
-					l.Address, l.HWAddress, l.Hostname, l.SubnetID, l.ValidLife, l.ExpireTime)
-			}
-		} else {
-			fmt.Printf("  %-40s %-20s %-15s %-12s %s\n", "Address", "DUID", "Hostname", "Lifetime", "Expires")
-			for _, l := range leases6 {
-				fmt.Printf("  %-40s %-20s %-15s %-12s %s\n",
-					l.Address, l.HWAddress, l.Hostname, l.ValidLife, l.ExpireTime)
-			}
-		}
-	}
-	return nil
-}
-
-// showDHCPDynamicDNS renders the DHCP dynamic-DNS config summary + runtime
-// counters (and, in detail mode, the owned records) for the in-process
-// interactive CLI (#1387 inc-2). Counters come from the daemon-injected
-// hooks; config comes from the active store.
-func (c *CLI) showDHCPDynamicDNS(detail bool) error {
-	cfg := c.store.ActiveConfig()
-	var ddns *config.DHCPDynamicDNSConfig
-	if cfg != nil {
-		ddns = cfg.System.DHCPServer.DynamicDNS
-	}
-	if ddns == nil {
-		fmt.Println("DHCP dynamic-DNS: not configured")
-		return nil
-	}
-	fmt.Println("DHCP Dynamic DNS:")
-	fmt.Printf("  Enabled:         %t\n", ddns.Enabled)
-	if ddns.Backend != "" {
-		fmt.Printf("  Backend:         %s\n", ddns.Backend)
-	}
-	if ddns.Domain != "" {
-		fmt.Printf("  Domain:          %s\n", ddns.Domain)
-	}
-	if ddns.UpdateServer != "" {
-		fmt.Printf("  Update server:   %s\n", ddns.UpdateServer)
-	}
-	if ddns.ConflictPolicy != "" {
-		fmt.Printf("  Conflict policy: %s\n", ddns.ConflictPolicy)
-	}
-	if ddns.TSIGKeyName != "" {
-		fmt.Printf("  TSIG key:        %s (secret redacted)\n", ddns.TSIGKeyName)
-	}
-
-	if c.ddnsStatsFn == nil {
-		fmt.Println("\n  Runtime counters: unavailable")
-		return nil
-	}
-	st := c.ddnsStatsFn()
-	if st == nil {
-		fmt.Println("\n  Runtime counters: unavailable (manager not running)")
-		return nil
-	}
-	if st.Degraded {
-		fmt.Printf("\n  ALARM: DDNS DEGRADED (fail-closed) — %s\n", st.DegradedReason)
-		fmt.Println("    Publishing and withdrawals are SUSPENDED until the ownership state is resolved.")
-	}
-	fmt.Println("\n  Counters:")
-	fmt.Printf("    Upserts:    ok=%d fail=%d\n", st.UpsertOK, st.UpsertFail)
-	fmt.Printf("    Deletes:    ok=%d fail=%d\n", st.DeleteOK, st.DeleteFail)
-	fmt.Printf("    Reconciles: ok=%d fail=%d\n", st.ReconcileOK, st.ReconcileFail)
-	fmt.Printf("    Skipped:    no-name=%d no-backend=%d conflict=%d ptr-notauth=%d\n",
-		st.SkippedNoName, st.SkippedNoBackend, st.SkippedConflict, st.SkippedPTRNotAuth)
-	fmt.Printf("    PTR deferred: %d total (lifetime), %d pending now\n",
-		st.PTRDeferred, st.PTRPendingNow)
-	fmt.Printf("    Owned records: %d\n", st.OwnedRecords)
-	if !st.LastReconcile.IsZero() {
-		fmt.Printf("    Last reconcile: %s (%d leases)\n",
-			st.LastReconcile.Format(time.RFC3339), st.LastReconcileN)
-	}
-
-	if detail && c.ddnsOwnedRecordsFn != nil {
-		recs := c.ddnsOwnedRecordsFn()
-		fmt.Println("\n  Owned records:")
-		if len(recs) == 0 {
-			fmt.Println("    none")
-		} else {
-			fmt.Printf("    %-32s %-6s %-39s %-26s %s\n", "FQDN", "Type", "Address", "PTR", "Pending")
-			for _, r := range recs {
-				pending := "-"
-				if r.PTRPending {
-					pending = "PTR"
-				}
-				fmt.Printf("    %-32s %-6s %-39s %-26s %s\n",
-					r.FQDN, r.ForwardType, r.Address, r.PTRName, pending)
-			}
-		}
-	}
-	return nil
-}
-
-func (c *CLI) showSNMP() error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil || cfg.System.SNMP == nil {
-		fmt.Println("No SNMP configured")
-		return nil
-	}
-	snmpCfg := cfg.System.SNMP
-
-	if snmpCfg.Location != "" {
-		fmt.Printf("Location:    %s\n", snmpCfg.Location)
-	}
-	if snmpCfg.Contact != "" {
-		fmt.Printf("Contact:     %s\n", snmpCfg.Contact)
-	}
-	if snmpCfg.Description != "" {
-		fmt.Printf("Description: %s\n", snmpCfg.Description)
-	}
-
-	if len(snmpCfg.Communities) > 0 {
-		fmt.Println("Communities:")
-		// #4111: mask the secret community name for any non-super-user login
-		// class (reuse the #4099/#4106 showConfigRedacted predicate). The
-		// authorization mode stays visible; only the community credential is
-		// masked. Super-user / unset class reads cleartext.
-		redactCommunity := c.showConfigRedacted()
-		for name, comm := range snmpCfg.Communities {
-			shown := name
-			if redactCommunity {
-				shown = config.SecretDataPlaceholder
-			}
-			fmt.Printf("  %s: %s\n", shown, comm.Authorization)
-		}
-	}
-
-	if len(snmpCfg.TrapGroups) > 0 {
-		fmt.Println("Trap groups:")
-		for name, tg := range snmpCfg.TrapGroups {
-			fmt.Printf("  %s: %s\n", name, strings.Join(tg.Targets, ", "))
-		}
-	}
-
-	if len(snmpCfg.V3Users) > 0 {
-		fmt.Println("SNMPv3 USM users:")
-		for name, u := range snmpCfg.V3Users {
-			auth := u.AuthProtocol
-			if auth == "" {
-				auth = "none"
-			}
-			priv := u.PrivProtocol
-			if priv == "" {
-				priv = "none"
-			}
-			fmt.Printf("  %s: auth=%s priv=%s\n", name, auth, priv)
-		}
-	}
-	return nil
-}
-
-func (c *CLI) showSNMPv3() error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil || cfg.System.SNMP == nil || len(cfg.System.SNMP.V3Users) == 0 {
-		fmt.Println("No SNMPv3 users configured")
-		return nil
-	}
-	fmt.Println("SNMPv3 USM Users:")
-	fmt.Printf("  %-20s %-12s %-12s\n", "User", "Auth", "Privacy")
-	for _, u := range cfg.System.SNMP.V3Users {
-		auth := u.AuthProtocol
-		if auth == "" {
-			auth = "none"
-		}
-		priv := u.PrivProtocol
-		if priv == "" {
-			priv = "none"
-		}
-		fmt.Printf("  %-20s %-12s %-12s\n", u.Name, auth, priv)
-	}
-	return nil
-}
-
-func (c *CLI) showLLDP() error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil || cfg.Protocols.LLDP == nil {
-		fmt.Println("LLDP not configured")
-		return nil
-	}
-	lldpCfg := cfg.Protocols.LLDP
-	if lldpCfg.Disable {
-		fmt.Println("LLDP: disabled")
-		return nil
-	}
-	fmt.Println("LLDP:")
-	interval := lldpCfg.Interval
-	if interval <= 0 {
-		interval = 30
-	}
-	holdMult := lldpCfg.HoldMultiplier
-	if holdMult <= 0 {
-		holdMult = 4
-	}
-	fmt.Printf("  Transmit interval: %ds\n", interval)
-	fmt.Printf("  Hold multiplier:   %d\n", holdMult)
-	fmt.Printf("  Hold time:         %ds\n", interval*holdMult)
-	if len(lldpCfg.Interfaces) > 0 {
-		var ifNames []string
-		for _, iface := range lldpCfg.Interfaces {
-			if iface.Disable {
-				ifNames = append(ifNames, iface.Name+" (disabled)")
-			} else {
-				ifNames = append(ifNames, iface.Name)
-			}
-		}
-		fmt.Printf("  Interfaces:        %s\n", strings.Join(ifNames, ", "))
-	}
-	if c.lldpNeighborsFn != nil {
-		neighbors := c.lldpNeighborsFn()
-		fmt.Printf("  Neighbors:         %d\n", len(neighbors))
-	}
-	return nil
-}
-
-func (c *CLI) showLLDPNeighbors() error {
-	if c.lldpNeighborsFn == nil {
-		fmt.Println("LLDP not running")
-		return nil
-	}
-	neighbors := c.lldpNeighborsFn()
-	if len(neighbors) == 0 {
-		fmt.Println("No LLDP neighbors discovered")
-		return nil
-	}
-	fmt.Printf("%-12s %-20s %-16s %-20s %-6s %s\n",
-		"Interface", "Chassis ID", "Port ID", "System Name", "TTL", "Age")
-	for _, n := range neighbors {
-		age := time.Since(n.LastSeen).Truncate(time.Second)
-		fmt.Printf("%-12s %-20s %-16s %-20s %-6d %s\n",
-			n.Interface, n.ChassisID, n.PortID, n.SystemName, n.TTL, age)
-	}
-	return nil
-}
-
-// showPortMirroring displays port mirroring (SPAN) configuration.
-
-func (c *CLI) showPortMirroring() error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("No active configuration")
-		return nil
-	}
-
-	pm := cfg.ForwardingOptions.PortMirroring
-	if pm == nil || len(pm.Instances) == 0 {
-		fmt.Println("No port-mirroring instances configured")
-		return nil
-	}
-
-	for name, inst := range pm.Instances {
-		fmt.Printf("Instance: %s\n", name)
-		if inst.InputRate > 0 {
-			fmt.Printf("  Input rate: 1/%d\n", inst.InputRate)
-		} else {
-			fmt.Printf("  Input rate: all packets\n")
-		}
-		if len(inst.Input) > 0 {
-			fmt.Printf("  Input interfaces: %s\n", strings.Join(inst.Input, ", "))
-		}
-		if inst.Output != "" {
-			fmt.Printf("  Output interface: %s\n", inst.Output)
 		}
 		fmt.Println()
 	}
